+++ learn-shop-admin-system/src/main/resources/processes/leave2.bpmn20.xml
@@ -1,0 +1,76 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:activiti="http://activiti.org/bpmn" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:omgdc="http://www.omg.org/spec/DD/20100524/DC" xmlns:omgdi="http://www.omg.org/spec/DD/20100524/DI" typeLanguage="http://www.w3.org/2001/XMLSchema" expressionLanguage="http://www.w3.org/1999/XPath" targetNamespace="http://www.activiti.org/processdef">
+  <process id="leave2" name="leave2" isExecutable="true">
+    <documentation>请假申请2</documentation>
+    <startEvent id="startEvent1"></startEvent>
+    <userTask id="apply" name="用户申请" activiti:formKey="apply_leave"></userTask>
+    <userTask id="approval" name="hr审核" activiti:candidateGroups="hr"></userTask>
+    <exclusiveGateway id="sid-6FE28C36-3802-4D3F-93E4-8A08CFA2DABC"></exclusiveGateway>
+    <endEvent id="sid-11CEF6D7-BDDF-432B-B7E4-C0DD0A281A3A">
+      <extensionElements>
+        <activiti:executionListener event="end" delegateExpression="${customExecutionListener}"></activiti:executionListener>
+      </extensionElements>
+    </endEvent>
+    <sequenceFlow id="_2" sourceRef="apply" targetRef="approval">
+      <conditionExpression xsi:type="tFormalExpression"><![CDATA[${transFlag=="1"}]]></conditionExpression>
+    </sequenceFlow>
+    <sequenceFlow id="_3" sourceRef="approval" targetRef="sid-6FE28C36-3802-4D3F-93E4-8A08CFA2DABC"></sequenceFlow>
+    <sequenceFlow id="_4" sourceRef="sid-6FE28C36-3802-4D3F-93E4-8A08CFA2DABC" targetRef="sid-11CEF6D7-BDDF-432B-B7E4-C0DD0A281A3A">
+      <conditionExpression xsi:type="tFormalExpression"><![CDATA[${transFlag=="1"}]]></conditionExpression>
+    </sequenceFlow>
+    <sequenceFlow id="_5" sourceRef="sid-6FE28C36-3802-4D3F-93E4-8A08CFA2DABC" targetRef="apply">
+      <conditionExpression xsi:type="tFormalExpression"><![CDATA[${transFlag=="0"}]]></conditionExpression>
+    </sequenceFlow>
+    <sequenceFlow id="_6" sourceRef="apply" targetRef="sid-11CEF6D7-BDDF-432B-B7E4-C0DD0A281A3A">
+      <conditionExpression xsi:type="tFormalExpression"><![CDATA[${transFlag=="0"}]]></conditionExpression>
+    </sequenceFlow>
+    <sequenceFlow id="_1" sourceRef="startEvent1" targetRef="apply"></sequenceFlow>
+  </process>
+  <bpmndi:BPMNDiagram id="BPMNDiagram_leave2">
+    <bpmndi:BPMNPlane bpmnElement="leave2" id="BPMNPlane_leave2">
+      <bpmndi:BPMNShape bpmnElement="startEvent1" id="BPMNShape_startEvent1">
+        <omgdc:Bounds height="30.0" width="30.0" x="90.0" y="160.0"></omgdc:Bounds>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape bpmnElement="apply" id="BPMNShape_apply">
+        <omgdc:Bounds height="80.0" width="100.0" x="210.0" y="134.0"></omgdc:Bounds>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape bpmnElement="approval" id="BPMNShape_approval">
+        <omgdc:Bounds height="80.0" width="100.0" x="420.0" y="134.0"></omgdc:Bounds>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape bpmnElement="sid-6FE28C36-3802-4D3F-93E4-8A08CFA2DABC" id="BPMNShape_sid-6FE28C36-3802-4D3F-93E4-8A08CFA2DABC">
+        <omgdc:Bounds height="40.0" width="40.0" x="615.0" y="153.0"></omgdc:Bounds>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape bpmnElement="sid-11CEF6D7-BDDF-432B-B7E4-C0DD0A281A3A" id="BPMNShape_sid-11CEF6D7-BDDF-432B-B7E4-C0DD0A281A3A">
+        <omgdc:Bounds height="28.0" width="28.0" x="735.0" y="159.0"></omgdc:Bounds>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNEdge bpmnElement="_1" id="BPMNEdge__1">
+        <omgdi:waypoint x="119.99968783492707" y="174.90322782041983"></omgdi:waypoint>
+        <omgdi:waypoint x="210.0" y="174.32258064516128"></omgdi:waypoint>
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge bpmnElement="_2" id="BPMNEdge__2">
+        <omgdi:waypoint x="310.0" y="174.0"></omgdi:waypoint>
+        <omgdi:waypoint x="420.0" y="174.0"></omgdi:waypoint>
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge bpmnElement="_3" id="BPMNEdge__3">
+        <omgdi:waypoint x="520.0" y="173.6969696969697"></omgdi:waypoint>
+        <omgdi:waypoint x="615.1204819277109" y="173.12048192771084"></omgdi:waypoint>
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge bpmnElement="_4" id="BPMNEdge__4">
+        <omgdi:waypoint x="654.5840707964602" y="173.41592920353983"></omgdi:waypoint>
+        <omgdi:waypoint x="735.0001358438572" y="173.06167341037948"></omgdi:waypoint>
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge bpmnElement="_5" id="BPMNEdge__5">
+        <omgdi:waypoint x="635.5" y="192.5"></omgdi:waypoint>
+        <omgdi:waypoint x="635.5" y="271.0"></omgdi:waypoint>
+        <omgdi:waypoint x="260.0" y="271.0"></omgdi:waypoint>
+        <omgdi:waypoint x="260.0" y="214.0"></omgdi:waypoint>
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge bpmnElement="_6" id="BPMNEdge__6">
+        <omgdi:waypoint x="260.0" y="134.0"></omgdi:waypoint>
+        <omgdi:waypoint x="260.0" y="89.0"></omgdi:waypoint>
+        <omgdi:waypoint x="749.0" y="89.0"></omgdi:waypoint>
+        <omgdi:waypoint x="749.0" y="159.0"></omgdi:waypoint>
+      </bpmndi:BPMNEdge>
+    </bpmndi:BPMNPlane>
+  </bpmndi:BPMNDiagram>
+</definitions>
