+++ learn-shop-ui-admin/src/views/process/apply/applyInfo/applyInfo.vue
@@ -3,13 +3,14 @@
     <template>
       <el-tabs v-model="activeName" @tab-click="handleClick">
         <el-tab-pane label="申请信息" name="first">
-          <!-- 请假申请 -->
-          <!--          <leave v-if="applyType === 'leave'"></leave>-->
+          <!-- 加载对应的页面 -->
           <component v-bind:is="applyType"></component>
         </el-tab-pane>
+        <!-- 流程图 -->
         <el-tab-pane label="申请流程" name="second" :lazy="true">
           <view-process-img></view-process-img>
         </el-tab-pane>
+        <!-- 批注信息 -->
         <el-tab-pane v-if="optionType !== 'add'" label="批注" name="third" :lazy="true">
           <comment-list></comment-list>
         </el-tab-pane>
@@ -19,9 +20,12 @@
 </template>
 
 <script>
+  // 申请页面---start 
   import leave from '../leave';
-  import CommentList from './commentList';
-  import ViewProcessImg from '../../ViewProcessImg';
+  import leave2 from '../leave2';
+  // 申请页面---end
+  import CommentList from './commentList'; // 批注信息
+  import ViewProcessImg from '../../ViewProcessImg'; // 流程图
 
   export default {
     name: "apply",
@@ -29,7 +33,8 @@
       CommentList,
       ViewProcessImg,
       CommentList,
-      leave
+      leave,
+      leave2
     },
     data() {
       return {
